more python 2to3 changes, build update

+++ model/mi_msi_model.py
@@ -85,7 +85,7 @@ class MSIModel(nn.Module):
             layer
         '''
         self.feature_extractor_part2 = nn.Sequential(
-            nn.Linear(64 * (self.coverage/2) * 10, self.num_features),
+            nn.Linear(64 * int(self.coverage/2) * 10, self.num_features),
             nn.ReLU(),
             nn.Linear(self.num_features, self.num_features),
             nn.ReLU(),
@@ -143,7 +143,7 @@ class MSIModel(nn.Module):
         out_4_2 = out_4_2 + res6
         final_instance_embed = self.relu(out_4_2)
 
-        final_instance_embed = final_instance_embed.view(-1, 64 * (self.coverage/2) * 10)
+        final_instance_embed = final_instance_embed.view(-1, 64 * int(self.coverage/2) * 10)
         I = self.feature_extractor_part2(final_instance_embed)  # N x num_features     
 
         # S is the sample-level embedding, which is the aggregation (via mean) of
